+++ src/memory.c
@@ -3,8 +3,8 @@
 
 int main(void) {
     struct RAM *maMemoire = init_RAM();
-    unsigned int *test;
-    test = malloc(sizeof(unsigned int));
+    uint8_t *test;
+    test = malloc(sizeof(uint8_t));
     write_RAM(maMemoire, 255, 12);
     read_RAM(maMemoire, 255, test);
     printf("%u\n", *test);
